requirements.txt
huggingface_hub==0.17.3
transformers==4.35.2
sentence-transformers==2.2.2
accelerate==0.24.1
datasets==2.14.7
auto-gptq==0.5.1
optimum==1.16.2
bitsandbytes==0.37.2
tiktoken==0.5.2
transformers_stream_generator==0.0.4
torch==2.0.1+cu126
torchvision==0.15.2+cu126
torchaudio==2.0.2
pandas==1.5.3
numpy==1.26.4
faiss-cpu==1.7.4
deep_translator==1.11.4
einops==0.7.0